issue #177 Set the algorithm of CPLEX
- CPLEX optimizer can be set to be deterministic or undeterministic. We currently set CPLEX to be deterministic (can be changed later).
- Clear model of CPLEX before solving another "instance" of the BIP problem.

+++ extensions/bip/src/edu/ucsc/dbtune/bip/interactions/CPlexInteraction.java
@@ -40,14 +40,15 @@ public class CPlexInteraction extends CPlexImplementer
                 double coef = 0.0;
                 Object found = mapVarCoef.get(var.getName());
                 if (found != null) 
-                    coef = ((Double)found).doubleValue();
+                    coef = (Double) found;
                       
                 listCoef[i] = coef;
             }
             cplex.addLe(cplex.scalProd(listCoef, vars), 0);
             
-            if (cplex.solve())
+            if (cplex.solve()) 
                 mapVariableValue = super.getMapVariableValue();
+            
             
         } catch (IloException e) {
             e.printStackTrace();
